+++ src/main/java/Params.java
@@ -4,7 +4,7 @@ import java.util.Properties;
 public class Params
 {
     public double gamma;
-    public int meanSpringLength;
+    public double meanSpringLength;
     public double spring;
     public int q2;
     public int width;
@@ -18,7 +18,7 @@ public class Params
         this.fis = new FileInputStream("src/main/resources/config");
         property.load(fis);
         this.gamma = Double.parseDouble(property.getProperty("gamma"));
-        this.meanSpringLength = Integer.parseInt(property.getProperty("meanSpringLength"));
+        this.meanSpringLength = Double.parseDouble(property.getProperty("meanSpringLength"));
         this.spring = Double.parseDouble(property.getProperty("spring"));
         this.q2 = Integer.parseInt(property.getProperty("q2"));
         this.width = Integer.parseInt(property.getProperty("width"));
